The actual path finding algorithm is complete, it does not consider the priority of the path yet, I'm not sure my prioritising it correctly as it has 2 disjoint costs for volume and weight. Currently priority of a route is volume_cost * weight_cost (with lower having a higher piority).
This brings about the scenario where there are two routes say A and B. Where A has weight cost (wc) of 10 and a volume cost (vc) of 1 and B has wc of 2 and vc of 10.

In this scenario, A has a lower priority (10 * 1)  than B (2 * 10). But if we have a package weighing 10 with volume 1, it is cheaper to send over B.
Cost for sending over A is 10*10 + 1*1 = 101
Cost for sending over B is 10*2 + 1*10 = 30